--- apps/roam/src/components/AdvancedNodeSearchDialog/advancedSearchSession.ts
@@ -1,7 +1,0 @@
-import type { SearchResult, SortConfig } from "./utils";
-
-export type AdvancedNodeSearchSession = {
-  query: string;
-  sort: SortConfig;
-  results: SearchResult[];
-};
