+++ apps/roam/src/utils/openDgSearchInSidebar.tsx
@@ -1,9 +1,42 @@
 import React from "react";
 import renderWithUnmount from "roamjs-components/util/renderWithUnmount";
-import { AdvancedSearchSidebarPanel } from "./AdvancedSearchSidebarPanel";
-import type { AdvancedNodeSearchSession } from "./advancedSearchSession";
+import { AdvancedSearchSidebarPanel } from "../components/AdvancedNodeSearchDialog/AdvancedSearchSidebarPanel";
+import getPageTitleByPageUid from "roamjs-components/queries/getPageTitleByPageUid";
+import openBlockInSidebar from "roamjs-components/writes/openBlockInSidebar";
+import {
+  SearchResult,
+  SortConfig,
+} from "~/components/AdvancedNodeSearchDialog/utils";
 
 const SIDEBAR_ROOT_ID = "dg-node-search-sidebar-root";
+
+export type AdvancedNodeSearchSession = {
+  query: string;
+  sort: SortConfig;
+  results: SearchResult[];
+};
+
+export const openSearchResultInMain = async (uid: string): Promise<void> => {
+  if (getPageTitleByPageUid(uid)) {
+    await window.roamAlphaAPI.ui.mainWindow.openPage({ page: { uid } });
+    return;
+  }
+  await window.roamAlphaAPI.ui.mainWindow.openBlock({ block: { uid } });
+};
+
+export const openSearchResultFromLinkEvent = async ({
+  uid,
+  shiftKey,
+}: {
+  uid: string;
+  shiftKey: boolean;
+}): Promise<void> => {
+  if (shiftKey) {
+    await openBlockInSidebar(uid);
+    return;
+  }
+  await openSearchResultInMain(uid);
+};
 
 let unmountSidebarSearch: (() => void) | null = null;
